backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
alembic==1.13.3
xgboost==2.1.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
shap==0.46.0
matplotlib==3.9.2
flwr==1.12.0
aiofiles==24.1.0
opencv-python==4.10.0.84
pydicom==2.4.4
